Add test for supplemental context to codeWhispererServer.test.ts
